refactor: 네이밍 수정과 함수분리(1)
