+++ NumberBaseball/NumberBaseball/UserNumberError.swift
@@ -1,0 +1,9 @@
+import Foundation
+
+enum userNumbersError: Error {
+    case noInput
+    case outOfRange
+    case wrongNumbersCount
+    case wrongFormat
+    case others
+}
